feat(bf16): merge BF16 scalar and vector tests into one suite
