--- tests/bf16-vector/Makefile
@@ -1,5 +1,0 @@
-NAME = bf16-vector-test
-SRCS =  main.c
-MARCH ?= rv64gcv_zba_zbb_zbc_zbs_zbkb_zbkc_zbkx_zknd_zkne_zknh_zkr_zksed_zksh_zkt_zfbfmin_zvfbfwma_zvfbfmin_zfhmin
-
-include $(AM_HOME)/Makefile.app
