+++ tests/bf16/Makefile
@@ -1,5 +1,5 @@
 NAME = bf16test
-SRCS =  main.c
-MARCH ?= rv64gc_zba_zbb_zbc_zbs_zbkb_zbkc_zbkx_zknd_zkne_zknh_zkr_zksed_zksh_zkt_zfbfmin
+SRCS = bf16.c test_scalar.c test_vec.c main.c
+MARCH ?= rv64gc_zba_zbb_zbc_zbs_zbkb_zbkc_zbkx_zknd_zkne_zknh_zkr_zksed_zksh_zkt_zfbfmin_zvfbfwma_zvfbfmin_zfhmin
 
 include $(AM_HOME)/Makefile.app
